+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 * Add section "Limitations" to README.md
 * Add and fix feature/test for tuple as trait field
 * Add feature/test for HashSet as trait field
+* Add feature/test for Customized Struct as trait field
 * Add feature/test for Dict(BtreeMap, similar to HashMap, but in order) as trait field
 
 ## 0.5.0 - March 7, 2024